UserCRUDProject/requirements.txt
asgiref==3.7.2
Django==4.2.2
django-phonenumber-field==7.1.0
djangorestframework==3.14.0
phonenumbers==8.13.15
Pillow==10.0.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
